* config.vcxproj, * configure_vc.awk, * doc/sdccman.lyx, * sdcc_vc.h.in, sdcc_vc_in.h: renamed sdcc_vc_in.h to sdcc_vc.h.in * src/SDCCval.c (byteOfVal), * src/z80/gen.c (genLeftShiftLiteral, genPointerGet), * src/mcs51/gen.c (genLeftShiftLiteral): fixed warnings * src/SDCCnaddr.hpp: renamed cfg_node to cfg_naddr_node * src/SDCCralloc.hpp: default_constructor_of_cfg_node_called and   cfg_node constructor disabled to fix bug 3521407 * src/SDCCtree_dec.hpp: set svn properties * support/regression/ports/mcs51-common/spec.mk: fixed SIM_TIMEOUT bug * support/regression/tests/bug-3495411.c: enabled for mcs51 * support/regression/tests/gcc-torture-execute-20000622-1.c, * support/regression/tests/gcc-torture-execute-20010924-1.c, * support/regression/tests/gcc-torture-execute-961004-1.c, * support/regression/tests/gcc-torture-execute-961125-1.c, * support/regression/tests/gcc-torture-execute-990404-1.c: suppressed warnings
git-svn-id: http://svn.code.sf.net/p/sdcc/code/trunk@7667 4a8a32a2-be11-0410-ad9d-d568d2c75423

+++ sdcc/src/SDCCval.c
@@ -1567,7 +1567,7 @@ byteOfVal (value * val, int offset)
 
   if (SPEC_NOUN (val->etype) == V_FLOAT)
     {
-      float f = SPEC_CVAL (val->etype).v_float;
+      float f = (float)SPEC_CVAL (val->etype).v_float;
 
       if (offset > 3)
         return 0;
@@ -1616,7 +1616,6 @@ byteOfVal (value * val, int offset)
   /* we are lost ! */
   werror (E_INTERNAL_ERROR, __FILE__, __LINE__, "byteOfVal: unknown value");
   return 0;
-  
 }
